Scope get event id endpoint to admins only (#817)
* Scope get event id endpoint to admins only

* Updated endpoint

* Updated interfaces

* Updated tests

* Generate swagger

* Regenerate schema and update endpoint in controller and tests

* Update tests to use admin jwt
